refactor(treewide): ✨ consolidate small files into logical homes
Merge four pairs of files that were split across too many units:

  path_reduction.go → paths.go
    WatchPathBuf and ReducePaths live with the rest of the path
    and filesystem abstractions.

  socket.go → communicate.go
    The Unix socket transport (SocketPath, Framing, connectClient)
    and the IPC protocol types are two halves of the same layer.

  nix_options.go → builder.go
    NixOptions is only used by the builder and daemon; 60 lines
    do not warrant a separate file.

  panic.go → main.go
    withPanicHandler, ExitError, and lorriVersion are the runtime
    infrastructure for the binary entry point.

default.nix srcs updated to match.
